% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anrpackageusingc-package.R
\docType{package}
\name{anrpackageusingc-package}
\alias{anrpackageusingc}
\alias{anrpackageusingc-package}
\title{anrpackageusingc: Demonstrates Calling 'C' Code}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dewey Dunnington \email{dewey@fishandwhistle.net} (\href{https://orcid.org/0000-0002-9415-4582}{ORCID})

}
\keyword{internal}
