% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-function.R
\name{my_function}
\alias{my_function}
\title{A Wrapper for a 'C' Function}
\usage{
my_function(input)
}
\arguments{
\item{input}{An input value to the C function}
}
\value{
The result of the C function.
}
\description{
A Wrapper for a 'C' Function
}
\examples{
my_function("some input")

}
